+++ test/cases/js-dynamic-import-js/expected/prefetched-delayed.ab6b.js
@@ -1,0 +1,1 @@
+"use strict";(self.webpackChunk=self.webpackChunk||[]).push([[104],{12:(e,u,d)=>{d.r(u),d.d(u,{default:()=>t});var c=d(151);function t(){(0,c.Z)("Executed: Prefetched delayed module!")}}}]);