+++ test/cases/js-dynamic-import-js/src/deferred.js
@@ -1,0 +1,5 @@
+import log from './log';
+
+export default function () {
+  log('Executed: Deferred module!');
+}