+++ test/cases/js-dynamic-import-js/src/log.js
@@ -1,0 +1,3 @@
+export default function (msg = 'Script executed.') {
+  console.log(msg);
+}