+++ test/cases/js-dynamic-import-js/src/prefetched.js
@@ -1,0 +1,5 @@
+import log from './log';
+
+export default function () {
+  log('Executed: Prefetched module!');
+}